refactor: move ExtensionDiagnosticsConfig to homeboy-extension-contract
ExtensionDiagnosticsConfig wraps Vec<ExtensionToolDiagnosticDeclaration> (its
element type already moved to homeboy-extension-contract by #8765). It's a pure
serde config type with a trivial is_empty() impl, so move it + the impl into
manifest_capability_config alongside its element type, and re-export from
extension::manifest to keep crate::extension::ExtensionDiagnosticsConfig call
sites (ExtensionManifest.diagnostics field, mod.rs re-export) unchanged.

Continues the extension-contract keystone campaign (#8757/#8760/#8764/#8765):
draining pure config types out of the extension subsystem so a future
homeboy-extension extraction depends on a slim seam. Reunites this config with
its element type that already lives in the contract.

Refs #8425
